fix: prevent deletion of default user address
This commit modifies the deletion process for user addresses, by preventing the deletion of an
address set as the default. It adds checks in both `delete.php` and `deleteAddressList.php` to
ensure that the address is not the default. If it is, an exception is thrown.

Additionally, locale files (in both German and English) have been modified to include the new
`exception.users.address.delete.default` string, which is used in the exception. This string
advises the user to set a different address as the default before attempting deletion.

+++ admin/ajax/users/address/delete.php
@@ -52,6 +52,19 @@ QUI::$Ajax->registerFunction(
 
         $User = QUI::getUsers()->get($uid);
         $Address = $User->getAddress($aid);
+        $StandardAddress = $User->getStandardAddress();
+
+        if (
+            $StandardAddress !== null
+            && $StandardAddress->getUUID() === $Address->getUUID()
+        ) {
+            throw new QUI\Exception(
+                QUI::getLocale()->get(
+                    'quiqqer/core',
+                    'exception.users.address.delete.default'
+                )
+            );
+        }
 
         $Address->delete();
     },
